Rough complete version, pre-office hours
Mostly filling out the psuedo code from first commit. Basically at this point the program reads in a list of processes from a text file, then "executes" all of them in a Round Robin fashion, giving each file either 25 cycles or until it completes. If I completely understood the requirements of this section of the project, this is a decent rough version of this part of the project. If office hours visit goes well then this only requires some minor tweaks to be good for turn in for the Round Robin portion of Part 1